requirements.txt
Django==4.2.13
djangorestframework==3.15.2
celery==5.3.6
redis==5.0.4
psycopg2-binary==2.9.12
django-environ==0.11.2
django-cors-headers==4.3.1
